requirements.txt
pandas
openpyxl
matplotlib
openai
python-dotenv
pyyaml>=6.0
numpy>=1.26
